requirements.txt
python-dotenv
fpdf
neo4j
fastapi
uvicorn
langgraph
langchain-google-genai
twilio
deepgram-sdk
neo4j
python-dotenv
websockets
streamlit
genai
generativeai
fastapi
uvicorn
requests
python-dotenv
pydantic
